requirements.txt
Flask==3.0.3
mysql-connector-python
MarkupSafe
Werkzeug
jinja2
bcrypt
blinker
click
itsdangerous
python-dotenv
PyJWT
